Rename _wp_clear_update_cache() to wp_clean_update_cache().
This makes it match most cache-clearing functions, including the ones it wraps. Also no need for it to be prefixed as "private."

wp_clean_plugins_cache() doesn't always exist, so as a quick fix, clear the transient we care to clear.

Merges [30856] to the 4.1 branch.

Fixes #30369.

Built from https://develop.svn.wordpress.org/branches/4.1@30870


git-svn-id: http://core.svn.wordpress.org/branches/4.1@30860 1a063a9b-81f0-0310-95a4-ce76da25c4cd