Remove PHP 7 from allowed failures list
PHP 7 has released it's 7th (and potentially final) Release Candidate today. While it is still not recommended for use in production, it is at a point that a test failure with it is something that needs to be triaged immediately. Congratulations to the PHP internals team on the upcoming final release of PHP 7.

Fixes #33410


Built from https://develop.svn.wordpress.org/trunk@35633


git-svn-id: http://core.svn.wordpress.org/trunk@35597 1a063a9b-81f0-0310-95a4-ce76da25c4cd