Customize: Make the menu edit button look like a link.
The button was introduced in [37901] to allow users switching to the selected menu for further edits. A link makes it more clear that the user is taken away from the current view.
This also adds an aria label and makes the button label more verbose, 'Edit Menu'.

Merge of [38189] to the 4.6 branch.

Props afercia.
Props helen for review.
See #36795.
Built from https://develop.svn.wordpress.org/branches/4.6@38190


git-svn-id: http://core.svn.wordpress.org/branches/4.6@38131 1a063a9b-81f0-0310-95a4-ce76da25c4cd