Build/Test Tools: Avoid skipping canonical tests that are connected to open Trac tickets.
This is a remnant of the test suite behaviour before #30284 fixed most of it and makes it far too easy to hide failing tests.

If any tests begin to fail as a result of this change then they should be fixed.

Fixes #40534

Built from https://develop.svn.wordpress.org/trunk@40518


git-svn-id: http://core.svn.wordpress.org/trunk@40394 1a063a9b-81f0-0310-95a4-ce76da25c4cd