Build tools: Clone and build from packages repo.
Instead of requiring the packages individually this commit changes it so the packages are used from the packages repository (gutenberg). This is a solution that makes updating the packages easier, as long as we don't have a monorepo to manage both WordPress & the JavaScript packages together.

Props youknowriad, gziolo, omarreiss, herregroen.
See #45145.

Built from https://develop.svn.wordpress.org/branches/5.0@43824


git-svn-id: http://core.svn.wordpress.org/branches/5.0@43653 1a063a9b-81f0-0310-95a4-ce76da25c4cd