Twenty Eleven: Correct `font-size` and button issues.
Update a mistake in Twenty Eleven's block editor styles that erroneously sets the `font-size` to `300`, when it should set the `font-weight` to `300`. This was causing browsers in quirks mode to make the editor font size very large.

Also, simplify some of the button block’s CSS selectors in the theme, to make sure the default colors don’t override the custom colors on the front end.

Props laurelfulford.

Merges [44190] and [44191] to trunk.

Fixes #45421, #45433.
Built from https://develop.svn.wordpress.org/trunk@44301


git-svn-id: http://core.svn.wordpress.org/trunk@44131 1a063a9b-81f0-0310-95a4-ce76da25c4cd