Posts, Post Types: Correctly show hierarchical post type hierarchy in admin.
In [44185], a bug was introduced where hierarchical post types would not display in the correct default order (hierarchically).

This was caused by a `! isset()` check, which returned `false` after [44185], causing the correct default value to not be applied. This switches that conditional to use an `empty()` check, ignoring the new empty string assignment that was added to prevent a PHP notice when `compact()` is called.

Merges [44338] to the 5.0 branch.

Props davidbinda.
Fixes #45711.
Built from https://develop.svn.wordpress.org/branches/5.0@44339


git-svn-id: http://core.svn.wordpress.org/branches/5.0@44169 1a063a9b-81f0-0310-95a4-ce76da25c4cd