Feeds: Ensure a feed’s `<lastBuildDate>` has the correct date.
The maximum modified time for a is currently being determined after an arbitrary date format is applied. Because the string sort in `max()` is being used to return the highest (most recent) modified value, the incorrect date is sometimes returned. This was introduced in [45247].

Merges [45367] to the 5.2 branch.

Reviewed by desrosj, SergeyBiryukov.
Props onlanka, Rarst, mukesh27.
Fixes #47304.
Built from https://develop.svn.wordpress.org/branches/5.2@45368


git-svn-id: http://core.svn.wordpress.org/branches/5.2@45179 1a063a9b-81f0-0310-95a4-ce76da25c4cd