Bootstrap/Load: Don't assume WP_CONTENT_DIR is defined
When the mysql extention isn't loaded and a custom db dropin is not in place, we give folks a nice error. However, we can't assume that the WP_CONTENT_DIR constant is set yet since this runs before we define default constants.

This fixes a PHP8 error.

See: #50913.


Built from https://develop.svn.wordpress.org/branches/5.5@49161


git-svn-id: http://core.svn.wordpress.org/branches/5.5@48923 1a063a9b-81f0-0310-95a4-ce76da25c4cd