Administration: Fix background color of plugin update row on small screens.
When the color change was done, the background of plugins with updates had to be manually changed. This section was missed, which caused a visual bug. This fix also removes some redundant CSS for the same element.

Follow-up to [50025].
Props afragen.
Fixes #52452.


Built from https://develop.svn.wordpress.org/trunk@50237


git-svn-id: http://core.svn.wordpress.org/trunk@49898 1a063a9b-81f0-0310-95a4-ce76da25c4cd