Robots: Remove contradictory directive check in `wp_robots()`.
Removes the mutually exclusive directives check in `wp_robots()`, ie allow both `follow` and `nofollow` to be specified and for `archive` and `noarchive` to be specified.

This fixes a bug in which WordPress would defer to the most permissive over the least permissive. When contradictory instructions are included, WordPress will defer to the search engine's or archivist's resolution policy: generally this is to observe the least, not most permissive.

Props Cybr, flixos90, SergeyBiryukov.
Merges [50566] to the 5.7 branch.
Fixes #52713.

Built from https://develop.svn.wordpress.org/branches/5.7@50587


git-svn-id: http://core.svn.wordpress.org/branches/5.7@50200 1a063a9b-81f0-0310-95a4-ce76da25c4cd