Build/Test Tools: Prevent PHPUnit tests on `push` for forks/private mirrors.
The reorganization of the PHPUnit workflow in [50441] unintentionally caused the tests to be run for every `push` event, even for forks and private mirrors.

Previously, the second job required the first one to pass, and the conditional check on the first prevented both from running. Because the first job is no longer required for the second, both jobs must have the appropriate conditional check.

Merges [50670] to the 5.6 branch.
Fixes #52983.
Built from https://develop.svn.wordpress.org/branches/5.6@50672


git-svn-id: http://core.svn.wordpress.org/branches/5.6@50284 1a063a9b-81f0-0310-95a4-ce76da25c4cd