Build/Test Tools: Add assertions to ensure version-controlled files are not modified during CI, and fix the `grunt clean` command.
Some tests and some commands that run during the build steps modify files that are under version control. This adds assertions to ensure that these files don't remain in a modified state after the test runs on CI.

This also fixes the `grunt clean` command which erroneously deletes `script-loader-packages.php`. This file is re-populated during the build, but deleting it during the clean is undesirable.

Merges [51355] to the 5.8 branch.

Fixes #53606

Built from https://develop.svn.wordpress.org/branches/5.8@51357


git-svn-id: http://core.svn.wordpress.org/branches/5.8@50966 1a063a9b-81f0-0310-95a4-ce76da25c4cd