Users: Allow any DB field to be returned by `WP_User_Query`.
Restore behaviour of `fields` parameter in `WP_User_Query` to allow developers to specify any database field to be returned either individually or as part of a subset. Add these fields to the documentation.

When a subset of `fields` includes the `id` paramater, include it in the results in both upper and lowercase to maintain backward compatibility.

Follow up to [53327].

Props dd32, pbearne, kraftbj, peterwilsoncc, SergeyBiryukov.
Merges [53362,53373] to the 6.0 branch.
Fixes #53177.

Built from https://develop.svn.wordpress.org/branches/6.0@53374


git-svn-id: http://core.svn.wordpress.org/branches/6.0@52963 1a063a9b-81f0-0310-95a4-ce76da25c4cd