Pings/Trackbacks: Remove a mention of the "Page" post type from the Discussion meta box.
This changeset removes a mention of the "Page" post type from the Discussion meta box as this is a generic meta box, used for multiple post types. It makes it more consistent with the other option available in the meta box ("Allow comments") and also with the corresponding option in the block editor.

Follow-up to [12323].

Props jeremyfelt, sabernhardt, audrasjb, virgar.
Fixes #57429.
See #11346.

Built from https://develop.svn.wordpress.org/trunk@55040


git-svn-id: http://core.svn.wordpress.org/trunk@54573 1a063a9b-81f0-0310-95a4-ce76da25c4cd